feat!: Update logging to use winston
BREAKING CHANGE:
The existing logging solution was a bit simple. With winston we can log
to multiple places (file, stdout etc) and different formats (like json).
This is ultimately better for deployments that will send their logs to
other services.

For any plugins, they will likely error on trying to log something.
Plugins will need to be updated so that they are aware of the newer
logger.